python/requirements.txt
pytest
pytest-asyncio
python-dotenv
aiohttp
python-dotenv
ijson